FAT-C358990 move the test script to a separate file
